Added function to transcript object assignement into Octave code. NEED TO ADD TRANSCIPT RULES, methods done.
Added functionlaity to read items from object assignement tree and to transcript them into some string. Still need to finish method to transcript 4 main elements for me:
    - Grid
    - Excitation
    - Port
    - Material
Each object is assigned to one of these 4 categories and these categories are applied onto object. So now when tree going through is finished is just needed to fill code lines which are generated based on tree how it's set by user in other words which object are assigned to which category.

Also need to implement function to save and load current setting into file.