Added 'conducting sheet' material support, now it's in Material tab, also for microstrip ports for Port tab, no code gneeration, added support into save/load witch try catch blocks so shouldn't affect .ini files where it's no defined. Prepared fo script generating implementation.
Have to add signals for material change, also lot of stuff to enable/disable UI items, save/load functions must to be modfiied, verified on some fileter design if it's running properly, at least there are no errors.
When materials removed which is used for microstrip that ports are automatically erased what is right dealing with them for now.